Assignment 03 Added Security Configuration, Different Users and Pagination

+++ src/main/java/com/cpan282/clotheswarehouse/model/Cloth.java
@@ -44,7 +44,7 @@ public class Cloth {
     @Min(2021)
     private int yearOfCreation;
     @Min(1000)
-    private BigDecimal price;
+    private double price;
     private Brand brandFrom;
     
     @Builder.Default
